src/PYTHON/requirements.txt
pyserial
csv
sys
